DRY up async exec functions
