+++ package.json
@@ -7,7 +7,8 @@
     "gitbook": "*"
   },
   "dependencies": {
-    "q": "1.0.1"
+    "q": "1.0.1",
+    "lodash": "3.10.1"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
